Properly configure dropbox and dist directory
Apparently the `~/.dropbox-dist` directory needs to be modified to be
non-read/writeable otherwise `systemctl` cannot properly determine the
state of the `dropbox` service. Modifying this directory allows
`dropbox` to properly start on login.

Also, there was a user `systemctl` target that needed to be removed to
prevent multiple versions of dropbox from starting and conflicting with
each other.