Move pacman config and reflector before installs
Pacman and the mirrorlist should be configured before packages are
installed. This moves all the pacman configuration, reflector
configuration and pacman cache updates into a separate file that is run
before packages are installed and the rest of the system is configured.

This also ensures that reflector updates the mirrorlist synchronously
the first time the system is configured.