Try some new key mappings
Since Caps Lock is Control/Escape the actual Control key can be used as
a "Switch Mode" key in Xmodmap. The switch mode can then be used to add
additional functionality to regular keys that are closer or on the home
row. I have moved all symbols down to the home row + Switch Mode and I
have moved the numbers down to the top row + Switch Mode.